refactor(P4-2): .service.ts サフィックスをハイフン区切りへ（Batch5a）
- ドット区切り *.service.ts → *-service.ts（applications/domains/infrastructures の
  18ファイル。assignee-gantt / task-dependency / scheduling / mapping / phase-* 等）
- import パスのみ更新（引用符境界で置換、export シンボル名は不変）

本番型チェッククリーン。
