+++ src/__integration_tests__/applications/wbs-sync/wbs-sync-application.integration.test.ts
@@ -1,6 +1,6 @@
 import { container } from '@/lib/inversify.config';
 import { SYMBOL } from '@/types/symbol';
-import { WbsSyncApplicationService } from '@/applications/wbs-sync/wbs-sync-application.service';
+import { WbsSyncApplicationService } from '@/applications/wbs-sync/wbs-sync-application-service';
 import { ExcelWbs } from '@/domains/sync/excel-wbs';
 import type { IExcelWbsRepository } from '@/applications/wbs-sync/iexcel-wbs-repository';
 import type { IWbsRepository } from '@/applications/wbs/iwbs-repository';
