+++ src/__tests__/domains/wbs/phase-coefficient-service.test.ts
@@ -1,4 +1,4 @@
-import { PhaseCoefficientService, PhaseHoursInput } from "@/domains/wbs/phase-coefficient.service";
+import { PhaseCoefficientService, PhaseHoursInput } from "@/domains/wbs/phase-coefficient-service";
 
 describe('PhaseCoefficientService', () => {
   describe('calculate', () => {
