+++ src/app/api/import-jobs/[id]/stream/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest } from 'next/server'
 import { container } from '@/lib/inversify.config'
 import { SYMBOL } from '@/types/symbol'
-import { IImportJobApplicationService } from '@/applications/import-job/import-job-application.service'
+import { IImportJobApplicationService } from '@/applications/import-job/import-job-application-service'
 
 interface Params {
   id: string
