+++ src/app/wbs/[id]/ganttv3/actions.ts
@@ -7,7 +7,7 @@ import { container } from "@/lib/inversify.config";
 import { Milestone, WbsTask } from "@/types/wbs";
 import { IWbsApplicationService } from "@/applications/wbs/wbs-application-service";
 import { IMilestoneApplicationService } from "@/applications/milestone/milestone-application-service";
-import type { ITaskDependencyService } from "@/applications/task-dependency/task-dependency.service";
+import type { ITaskDependencyService } from "@/applications/task-dependency/task-dependency-service";
 import { DependencyType } from "@/components/ganttv3/gantt";
 import { statusColor } from "@/components/ganttv3/utils/taskFormat";
 import { ProgressMeasurementMethod } from "@/types/progress-measurement";
