+++ src/applications/schedule-generator/schedule-generate-service.ts
