+++ src/domains/forecast/forecast-calculation-service.ts
